+++ Chat/src/Messages/ClientRequestUpdateMessage.java
@@ -7,7 +7,7 @@ public class ClientRequestUpdateMessage extends Message {
 	public static final long minSize = 143;
 
 	public ClientRequestUpdateMessage(int _op, long _length, long _reserved,
-			String _options, int[] body) {
+			String _options, byte[] body) {
 		super(_op, _length, _reserved, _options);
 		processBody(body);
 		if (op != 6) {
@@ -25,12 +25,12 @@ public class ClientRequestUpdateMessage extends Message {
 		}
 	}
 
-	private void processBody(int[] body) {
+	private void processBody(byte[] body) {
 		if (body.length != 143) {
 			correct = false;
 			return;
 		}
-		int[] senderUserArray = new int[128];
+		byte[] senderUserArray = new byte[128];
 		for (int i = 0; i < body.length; i++) {
 			senderUserArray[i] = body[i];
 		}
